Create a interface for graphic render to maps
