+++ src/main/java/org/rpgrunner/controller/MapController.java
@@ -7,13 +7,13 @@ import org.rpgrunner.character.CharacterAnimation;
 import org.rpgrunner.character.CharacterElement;
 import org.rpgrunner.character.movement.MovementCommand;
 import org.rpgrunner.character.movement.PlayerMovement;
-import org.rpgrunner.graphics.GraphicsRender;
+import org.rpgrunner.graphics.MapGraphicsRender;
 import org.rpgrunner.helper.Camera;
 import org.rpgrunner.helper.MapHelper;
 import org.rpgrunner.map.Map;
 
 public class MapController implements Controller {
-    private final GraphicsRender graphicsRender;
+    private final MapGraphicsRender mapGraphicsRender;
     private final Camera camera;
     private final Vector characterElements;
     private final MapHelper mapHelper;
@@ -22,11 +22,11 @@ public class MapController implements Controller {
     private PlayerMovement playerMovement;
 
     public MapController(
-        final GraphicsRender gameGraphicsRender,
+        final MapGraphicsRender gameGraphicsRender,
         final Camera gameCamera,
         final MapHelper gameMapHelper
     ) {
-        graphicsRender = gameGraphicsRender;
+        mapGraphicsRender = gameGraphicsRender;
         camera = gameCamera;
         characterElements = new Vector(1);
         mapHelper = gameMapHelper;
@@ -38,7 +38,7 @@ public class MapController implements Controller {
 
         mapHelper.setMap(map);
         camera.setMap(map);
-        graphicsRender.setMap(map);
+        mapGraphicsRender.setMap(map);
         removeAllNPCs();
     }
 
@@ -49,7 +49,7 @@ public class MapController implements Controller {
             characterElements.addElement(playerCharacterElement);
         }
 
-        graphicsRender.setCharacterElements(characterElements);
+        mapGraphicsRender.setCharacterElements(characterElements);
     }
 
     public Map getMap() {
@@ -78,7 +78,7 @@ public class MapController implements Controller {
     }
 
     public void render() {
-        graphicsRender.render();
+        mapGraphicsRender.render();
     }
 
     private void executeMovementCommand(
@@ -123,11 +123,11 @@ public class MapController implements Controller {
         final CharacterElement characterElement
     ) {
         characterElements.removeElement(characterElement);
-        graphicsRender.setCharacterElements(characterElements);
+        mapGraphicsRender.setCharacterElements(characterElements);
     }
 
     public void addCharacterElement(final CharacterElement characterElement) {
         characterElements.addElement(characterElement);
-        graphicsRender.setCharacterElements(characterElements);
+        mapGraphicsRender.setCharacterElements(characterElements);
     }
 }
