+++ src/main/java/org/rpgrunner/graphics/GraphicsRender.java
@@ -1,12 +1,6 @@
 package org.rpgrunner.graphics;
 
-import java.util.Vector;
-
-import org.rpgrunner.map.Map;
-
 public interface GraphicsRender {
-    void setMap(Map map);
-    void setCharacterElements(Vector characterElements);
     void showMessage(String message);
     void hideMessage();
     void render();
