+++ src/main/java/org/rpgrunner/graphics/MapGraphicsRender.java
@@ -1,0 +1,10 @@
+package org.rpgrunner.graphics;
+
+import java.util.Vector;
+
+import org.rpgrunner.map.Map;
+
+public interface MapGraphicsRender extends GraphicsRender {
+    void setMap(Map map);
+    void setCharacterElements(Vector characterElements);
+}
