+++ src/test/java/org/rpgrunner/test/mock/graphics/GraphicsRenderSpy.java
@@ -1,13 +1,8 @@
 package org.rpgrunner.test.mock.graphics;
 
-import java.util.Vector;
-
 import org.rpgrunner.graphics.GraphicsRender;
-import org.rpgrunner.map.Map;
 
 public class GraphicsRenderSpy implements GraphicsRender {
-    private Map map;
-    private Vector characterElements;
     private boolean renderCalled;
     private String lastMessage;
     private boolean showingMessage;
@@ -15,22 +10,6 @@ public class GraphicsRenderSpy implements GraphicsRender {
     public GraphicsRenderSpy() {
         renderCalled = false;
         showingMessage = false;
-    }
-
-    public void setMap(final Map newMap) {
-        map = newMap;
-    }
-
-    public Map getMap() {
-        return map;
-    }
-
-    public void setCharacterElements(final Vector newCharacterElements) {
-        characterElements = newCharacterElements;
-    }
-
-    public Vector getCharacterElements() {
-        return characterElements;
     }
 
     public void showMessage(final String message) {
